Backend/blip_service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
transformers==4.55.4
torch>=2.2.0
pillow==11.3.0
python-multipart==0.0.20
